+++ packages/codegen-ui-react/lib/react-studio-template-renderer-helper.ts
@@ -38,7 +38,6 @@ import ts, {
   createProgram,
   BindingName,
   Expression,
-  PropertyAssignment,
   ArrowFunction,
   CallExpression,
   Identifier,
@@ -292,105 +291,6 @@ export const buildBaseCollectionVariableStatement = (identifier: string | Bindin
           undefined,
           undefined,
           factory.createPropertyAccessExpression(expression, factory.createIdentifier('items')),
-        ),
-      ],
-      ts.NodeFlags.Const,
-    ),
-  );
-};
-/*
-  { [key]: relatedModel.filter(model => model.{relatedField} === item.id) }
- */
-export const buildPropAssignmentWithFilter = (key: string, relatedModel: string, relatedField: string) => {
-  return factory.createPropertyAssignment(
-    factory.createIdentifier(key),
-    factory.createCallExpression(
-      factory.createPropertyAccessExpression(
-        factory.createIdentifier(relatedModel),
-        factory.createIdentifier('filter'),
-      ),
-      undefined,
-      [
-        factory.createArrowFunction(
-          undefined,
-          undefined,
-          [
-            factory.createParameterDeclaration(
-              undefined,
-              undefined,
-              undefined,
-              factory.createIdentifier('model'),
-              undefined,
-              undefined,
-              undefined,
-            ),
-          ],
-          undefined,
-          factory.createToken(ts.SyntaxKind.EqualsGreaterThanToken),
-          factory.createBinaryExpression(
-            factory.createPropertyAccessExpression(
-              factory.createIdentifier('model'),
-              factory.createIdentifier(relatedField),
-            ),
-            factory.createToken(ts.SyntaxKind.EqualsEqualsEqualsToken),
-            // TODO: Change this once datastore supports custom primary keys
-            factory.createPropertyAccessExpression(factory.createIdentifier('item'), factory.createIdentifier('id')),
-          ),
-        ),
-      ],
-    ),
-  );
-};
-
-/*
-const {variableName} = {collectionCall}
-  .items.map((item) => ({ ...item, {propertyAssingments[]} }))
-*/
-export const buildCollectionWithItemMap = (
-  variableName: string,
-  collectionCall: Expression,
-  propAssignments: PropertyAssignment[],
-) => {
-  return factory.createVariableStatement(
-    undefined,
-    factory.createVariableDeclarationList(
-      [
-        factory.createVariableDeclaration(
-          factory.createIdentifier(variableName),
-          undefined,
-          undefined,
-          factory.createCallExpression(
-            factory.createPropertyAccessExpression(
-              factory.createPropertyAccessExpression(collectionCall, factory.createIdentifier('items')),
-              factory.createIdentifier('map'),
-            ),
-            undefined,
-            [
-              factory.createArrowFunction(
-                undefined,
-                undefined,
-                [
-                  factory.createParameterDeclaration(
-                    undefined,
-                    undefined,
-                    undefined,
-                    factory.createIdentifier('item'),
-                    undefined,
-                    undefined,
-                    undefined,
-                  ),
-                ],
-                undefined,
-                factory.createToken(ts.SyntaxKind.EqualsGreaterThanToken),
-                factory.createParenthesizedExpression(
-                  factory.createObjectLiteralExpression(
-                    [factory.createSpreadAssignment(factory.createIdentifier('item')), ...propAssignments],
-                    false,
-                  ),
-                ),
-              ),
-            ],
-          ),
         ),
       ],
       ts.NodeFlags.Const,