+++ packages/codegen-ui-golden-files/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.6.0](https://github.com/aws-amplify/amplify-codegen-ui/compare/v2.5.8...v2.6.0) (2022-12-14)
+
+**Note:** Version bump only for package @aws-amplify/codegen-ui-golden-files
+
+## [2.5.8](https://github.com/aws-amplify/amplify-codegen-ui/compare/v2.5.7...v2.5.8) (2022-12-08)
+
+**Note:** Version bump only for package @aws-amplify/codegen-ui-golden-files
+
+## [2.5.7](https://github.com/aws-amplify/amplify-codegen-ui/compare/v2.5.6...v2.5.7) (2022-12-06)
+
+**Note:** Version bump only for package @aws-amplify/codegen-ui-golden-files
+
 ## [2.5.6](https://github.com/aws-amplify/amplify-codegen-ui/compare/v2.5.5...v2.5.6) (2022-11-23)
 
 **Note:** Version bump only for package @aws-amplify/codegen-ui-golden-files